Move foreign-memory kernels into SafeRE MR-JAR
Refs #692
