--- safere-foreign/pom.xml
@@ -1,102 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
-                             http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.safere</groupId>
-    <artifactId>safere-parent</artifactId>
-    <version>0.11.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>safere-foreign</artifactId>
-  <packaging>jar</packaging>
-
-  <name>SafeRE Foreign Memory</name>
-  <description>Foreign Memory (MemorySegment) scanning kernels for SafeRE</description>
-
-  <properties>
-    <maven.compiler.release>22</maven.compiler.release>
-  </properties>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.junit</groupId>
-        <artifactId>junit-bom</artifactId>
-        <version>${junit.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.safere</groupId>
-      <artifactId>safere</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.assertj</groupId>
-      <artifactId>assertj-core</artifactId>
-      <version>3.27.3</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <configuration>
-          <compilerArgs combine.children="append">
-            <arg>--add-modules</arg>
-            <arg>jdk.incubator.vector</arg>
-            <arg>--add-reads</arg>
-            <arg>org.safere=jdk.incubator.vector</arg>
-            <arg>--add-reads</arg>
-            <arg>org.safere.foreign=jdk.incubator.vector</arg>
-            <arg>-Xlint:-incubating</arg>
-          </compilerArgs>
-        </configuration>
-        <executions>
-          <execution>
-            <id>default-testCompile</id>
-            <phase>test-compile</phase>
-            <goals>
-              <goal>testCompile</goal>
-            </goals>
-            <configuration>
-              <compilerArgs combine.children="append">
-                <arg>--add-modules</arg>
-                <arg>jdk.incubator.vector</arg>
-                <arg>--add-reads</arg>
-                <arg>org.safere=jdk.incubator.vector</arg>
-                <arg>--add-reads</arg>
-                <arg>org.safere.foreign=jdk.incubator.vector</arg>
-                <arg>-Xlint:-incubating</arg>
-              </compilerArgs>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <version>3.5.6</version>
-        <configuration>
-          <argLine>--add-modules=jdk.incubator.vector --add-reads org.safere=jdk.incubator.vector --add-reads org.safere.foreign=jdk.incubator.vector</argLine>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-</project>
