--- safere-foreign/src/main/java/module-info.java
@@ -1,12 +1,0 @@
-// This file is part of a Java port of RE2 (https://github.com/google/re2).
-// Original RE2 code is Copyright (c) 2009 The RE2 Authors.
-// Modifications and Java port Copyright (c) 2026 Eddie Aftandilian.
-// Licensed under the BSD 3-Clause License (see LICENSE file).
-
-/** SafeRE Foreign Memory: MemorySegment scanning kernels for SafeRE. */
-module org.safere.foreign {
-  requires org.safere;
-  requires static jdk.incubator.vector;
-
-  exports org.safere.foreign;
-}
