+++ safere/pom.xml
@@ -172,7 +172,9 @@ final class WorkCounterConfig {
           <java>
             <includes>
               <include>src/main/java/**/*.java</include>
+              <include>src/main/java-jdk22/**/*.java</include>
               <include>src/test/java/**/*.java</include>
+              <include>src/test/java-jdk22/**/*.java</include>
             </includes>
           </java>
         </configuration>
@@ -185,6 +187,7 @@ final class WorkCounterConfig {
           <archive>
             <manifestEntries>
               <Automatic-Module-Name>org.safere</Automatic-Module-Name>
+              <Multi-Release>true</Multi-Release>
             </manifestEntries>
           </archive>
         </configuration>
@@ -272,6 +275,127 @@ final class WorkCounterConfig {
   </build>
 
   <profiles>
+    <profile>
+      <id>jdk22-foreign-memory</id>
+      <activation>
+        <jdk>[22,)</jdk>
+      </activation>
+      <build>
+        <plugins>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-compiler-plugin</artifactId>
+            <executions>
+              <execution>
+                <id>compile-jdk22-foreign-memory</id>
+                <phase>process-classes</phase>
+                <goals>
+                  <goal>compile</goal>
+                </goals>
+                <configuration>
+                  <release>22</release>
+                  <compileSourceRoots combine.self="override">
+                    <compileSourceRoot>${project.basedir}/src/main/java-jdk22</compileSourceRoot>
+                  </compileSourceRoots>
+                  <multiReleaseOutput>true</multiReleaseOutput>
+                  <compilerArgs combine.self="override">
+                    <arg>--add-modules</arg>
+                    <arg>jdk.incubator.vector</arg>
+                    <arg>--patch-module</arg>
+                    <arg>org.safere=${project.build.outputDirectory}</arg>
+                    <arg>-Xlint:-incubating</arg>
+                  </compilerArgs>
+                  <annotationProcessorPaths combine.self="override"/>
+                </configuration>
+              </execution>
+              <execution>
+                <id>compile-jdk22-foreign-memory-tests</id>
+                <phase>test-compile</phase>
+                <goals>
+                  <goal>testCompile</goal>
+                </goals>
+                <configuration>
+                  <release>22</release>
+                  <outputDirectory>${project.build.directory}/test-classes-jdk22</outputDirectory>
+                  <compileSourceRoots combine.self="override">
+                    <compileSourceRoot>${project.basedir}/src/test/java-jdk22</compileSourceRoot>
+                    <!-- Make the versioned implementation visible while compiling its tests. -->
+                    <compileSourceRoot>
+                      ${project.build.outputDirectory}/META-INF/versions/22
+                    </compileSourceRoot>
+                  </compileSourceRoots>
+                  <compilerArgs combine.self="override">
+                    <arg>--add-modules</arg>
+                    <arg>jdk.incubator.vector</arg>
+                    <arg>-Xlint</arg>
+                    <arg>-Xlint:-incubating</arg>
+                    <arg>-Werror</arg>
+                    <arg>-XDcompilePolicy=simple</arg>
+                    <arg>--should-stop=ifError=FLOW</arg>
+                    <arg>-Xplugin:ErrorProne -XepCompilingTestOnlyCode ${error-prone.checks}</arg>
+                  </compilerArgs>
+                </configuration>
+              </execution>
+            </executions>
+          </plugin>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-antrun-plugin</artifactId>
+            <executions>
+              <execution>
+                <id>package-jdk22-test-jar</id>
+                <phase>process-test-classes</phase>
+                <goals>
+                  <goal>run</goal>
+                </goals>
+                <configuration>
+                  <target>
+                    <jar
+                        basedir="${project.build.outputDirectory}"
+                        destfile="${project.build.directory}/${project.build.finalName}-jdk22-tests.jar">
+                      <manifest>
+                        <attribute name="Automatic-Module-Name" value="org.safere"/>
+                        <attribute name="Multi-Release" value="true"/>
+                      </manifest>
+                    </jar>
+                  </target>
+                </configuration>
+              </execution>
+            </executions>
+          </plugin>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-surefire-plugin</artifactId>
+            <configuration>
+              <argLine>
+                @{argLine} --add-modules=jdk.incubator.vector
+                --add-reads org.safere=jdk.incubator.vector
+              </argLine>
+              <classesDirectory>
+                ${project.build.directory}/${project.build.finalName}-jdk22-tests.jar
+              </classesDirectory>
+            </configuration>
+            <executions>
+              <execution>
+                <id>test-jdk22-foreign-memory</id>
+                <phase>test</phase>
+                <goals>
+                  <goal>test</goal>
+                </goals>
+                <configuration>
+                  <testClassesDirectory>
+                    ${project.build.directory}/test-classes-jdk22
+                  </testClassesDirectory>
+                  <includes>
+                    <include>**/MemorySegmentScanEquivalenceTest.class</include>
+                  </includes>
+                </configuration>
+              </execution>
+            </executions>
+          </plugin>
+        </plugins>
+      </build>
+    </profile>
     <profile>
       <id>vector-tests</id>
       <build>
