+++ safere/src/main/java-jdk22/org/safere/SegmentByteVectorScan.java
@@ -3,7 +3,7 @@
 // Modifications and Java port Copyright (c) 2026 Eddie Aftandilian.
 // Licensed under the BSD 3-Clause License (see LICENSE file).
 
-package org.safere.foreign;
+package org.safere;
 
 import static org.safere.internal.Ascii.toLowerCase;
 import static org.safere.internal.Ascii.toUpperCase;
@@ -19,7 +19,7 @@ import jdk.incubator.vector.VectorSpecies;
 import org.safere.internal.Swar;
 
 /** Stateless 1-byte SIMD scanning kernels over {@link MemorySegment}. */
-public final class SegmentByteVectorScan {
+final class SegmentByteVectorScan {
 
   private static final VectorSpecies<Byte> SPECIES = ByteVector.SPECIES_PREFERRED;
   private static final ByteOrder NATIVE_ORDER = ByteOrder.nativeOrder();
