+++ safere/src/main/java/module-info.java
@@ -6,5 +6,4 @@
 /** SafeRE: a linear-time regular expression matching library for Java. */
 module org.safere {
   exports org.safere;
-  exports org.safere.internal;
 }
